+++ app/src/main/java/com/jin/honey/feature/home/ui/content/HomeReviewRanking.kt
@@ -49,7 +49,7 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import coil.compose.AsyncImage
 import com.jin.honey.R
-import com.jin.domain.model.review.ReviewRankPreview
+import com.jin.domain.review.ReviewRankPreview
 import com.jin.ui.theme.PointColor
 import com.jin.ui.theme.ReviewRankingBoxBackgroundColor
 import com.jin.ui.theme.ReviewRankingBoxBorderColor
