+++ app/src/main/java/com/jin/honey/feature/review/ui/ReviewViewModel.kt
@@ -4,7 +4,7 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.jin.feature.ui.state.UiState
 import com.jin.domain.usecase.GetReviewWithIngredientUseCase
-import com.jin.domain.model.review.ReviewPreview
+import com.jin.domain.review.ReviewPreview
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.launch
@@ -13,7 +13,7 @@ class ReviewViewModel(
     private val getReviewWithIngredientUseCase: GetReviewWithIngredientUseCase,
 ) : ViewModel() {
     private val _reviewsState = MutableStateFlow<UiState<List<ReviewPreview>>>(UiState.Loading)
-    val reviewsState: StateFlow<UiState<List< ReviewPreview>>> = _reviewsState
+    val reviewsState: StateFlow<UiState<List<ReviewPreview>>> = _reviewsState
 
     fun fetchReview(menuName: String) {
         viewModelScope.launch {
