+++ app/src/main/java/com/jin/honey/feature/reviewwrite/ui/ReviewWriteScreen.kt
@@ -34,8 +34,8 @@ import com.jin.feature.ui.state.UiState
 import com.jin.honey.R
 import com.jin.honey.feature.reviewwrite.ui.content.MenuReviewWriteScreen
 import com.jin.domain.food.model.CategoryType
-import com.jin.domain.model.review.Review
-import com.jin.domain.model.review.ReviewContent
+import com.jin.domain.review.Review
+import com.jin.domain.review.ReviewContent
 import kotlinx.coroutines.launch
 import java.time.Instant
 import java.time.LocalDate
