+++ app/src/main/java/com/jin/honey/feature/reviewwrite/ui/ReviewWriteViewModel.kt
@@ -7,7 +7,7 @@ import com.jin.feature.ui.state.UiState
 import com.jin.domain.order.model.Order
 import com.jin.domain.usecase.GetOrderDetailUseCase
 import com.jin.domain.usecase.WriteReviewUseCase
-import com.jin.domain.model.review.Review
+import com.jin.domain.review.Review
 import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
