+++ data/src/main/java/com/jin/data/firestore/FireStoreDataSource.kt
@@ -1,7 +1,7 @@
 package com.jin.data.firestore
 
 import com.jin.domain.food.model.Food
-import com.jin.domain.model.review.Review
+import com.jin.domain.review.Review
 import com.jin.domain.recipe.model.Recipe
 
 interface FireStoreDataSource {
