+++ data/src/main/java/com/jin/data/review/ReviewRepositoryImpl.kt
@@ -4,10 +4,10 @@ import android.util.Log
 import com.jin.database.datasource.ReviewTrackingDataSource
 import com.jin.database.entities.ReviewEntity
 import com.jin.data.firestore.FireStoreDataSource
-import com.jin.domain.repositories.ReviewRepository
+import com.jin.domain.review.ReviewRepository
 import com.jin.domain.food.model.CategoryType
-import com.jin.domain.model.review.Review
-import com.jin.domain.model.review.ReviewContent
+import com.jin.domain.review.Review
+import com.jin.domain.review.ReviewContent
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
 import java.time.Instant
@@ -25,7 +25,7 @@ class ReviewRepositoryImpl(
         }
     }
 
-    override suspend fun writtenReviewSave(reviews: List< Review>): Result<Unit> {
+    override suspend fun writtenReviewSave(reviews: List<Review>): Result<Unit> {
         return try {
             withContext(Dispatchers.IO) {
                 for (review in reviews) {
@@ -38,11 +38,11 @@ class ReviewRepositoryImpl(
         }
     }
 
-    override suspend fun fetchMenuReview(menuName: String): List< Review> {
+    override suspend fun fetchMenuReview(menuName: String): List<Review> {
         return try {
             withContext(Dispatchers.IO) {
                 val entities = db.queryOnlyOneMenuReview(menuName)
-                val reviews = mutableListOf< Review>()
+                val reviews = mutableListOf<Review>()
                 for (entity in entities) {
                     reviews.add(entity.toDomain())
                 }
@@ -53,11 +53,11 @@ class ReviewRepositoryImpl(
         }
     }
 
-    override suspend fun fetchReviewByCategory(categoryName: String): List< Review> {
+    override suspend fun fetchReviewByCategory(categoryName: String): List<Review> {
         return try {
             withContext(Dispatchers.IO) {
                 val entities = db.queryReviewByCategory(categoryName)
-                val reviews = mutableListOf< Review>()
+                val reviews = mutableListOf<Review>()
                 for (entity in entities) {
                     reviews.add(entity.toDomain())
                 }
@@ -68,7 +68,7 @@ class ReviewRepositoryImpl(
         }
     }
 
-    private fun  Review.toEntity(): ReviewEntity {
+    private fun Review.toEntity(): ReviewEntity {
         return ReviewEntity(
             id = id ?: 0,
             orderKey = orderKey,
