--- domain/src/main/java/com/jin/domain/repositories/ReviewRepository.kt
@@ -1,10 +1,0 @@
-package com.jin.domain.repositories
-
-import com.jin.domain.model.review.Review
-
-interface ReviewRepository {
-    suspend fun syncReviews()
-    suspend fun writtenReviewSave(reviews: List<Review>): Result<Unit>
-    suspend fun fetchMenuReview(menuName: String): List< Review>
-    suspend fun fetchReviewByCategory(categoryName: String): List< Review>
-}
