+++ domain/src/main/java/com/jin/domain/review/ReviewRepository.kt
@@ -1,0 +1,8 @@
+package com.jin.domain.review
+
+interface ReviewRepository {
+    suspend fun syncReviews()
+    suspend fun writtenReviewSave(reviews: List<Review>): Result<Unit>
+    suspend fun fetchMenuReview(menuName: String): List<Review>
+    suspend fun fetchReviewByCategory(categoryName: String): List<Review>
+}
