+++ domain/src/main/java/com/jin/domain/usecase/GetReviewUseCase.kt
@@ -1,7 +1,7 @@
 package com.jin.domain.usecase
 
-import com.jin.domain.repositories.ReviewRepository
-import com.jin.domain.model.review.Review
+import com.jin.domain.review.ReviewRepository
+import com.jin.domain.review.Review
 
 class GetReviewUseCase(private val repository: ReviewRepository) {
     suspend operator fun invoke(menuName: String): Result<List<Review>> {
