+++ domain/src/main/java/com/jin/domain/usecase/GetReviewWithIngredientUseCase.kt
@@ -1,15 +1,15 @@
 package com.jin.domain.usecase
 
 import com.jin.domain.order.OrderRepository
-import com.jin.domain.repositories.ReviewRepository
-import com.jin.domain.model.review.ReviewPreview
+import com.jin.domain.review.ReviewRepository
+import com.jin.domain.review.ReviewPreview
 
 class GetReviewWithIngredientUseCase(
     private val repository: ReviewRepository,
     private val orderRepository: OrderRepository
 ) {
     suspend operator fun invoke(menuName: String): Result<List<ReviewPreview>> {
-        val reviewPreviews = mutableListOf< ReviewPreview>()
+        val reviewPreviews = mutableListOf<ReviewPreview>()
         val reviews = repository.fetchMenuReview(menuName)
         for (review in reviews) {
             val cart = orderRepository.fetchOrderIngredients(review.orderKey, review.menuName)
