+++ domain/src/main/java/com/jin/domain/usecase/SyncReviewsUseCase.kt
@@ -1,6 +1,6 @@
 package com.jin.domain.usecase
 
-import com.jin.domain.repositories.ReviewRepository
+import com.jin.domain.review.ReviewRepository
 
 class SyncReviewsUseCase(private val repository: ReviewRepository) {
     suspend operator fun invoke() {
